% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_method.R
\name{combat_type}
\alias{combat_type}
\title{Create a ComBat Type Object}
\usage{
combat_type(type)
}
\arguments{
\item{type}{Character string specifying the ComBat implementation type,
such as \code{"univariate"}.}
}
\value{
An object whose class is set to \code{type}.
}
\description{
Creates an object used for S3 dispatch within the ComBat estimation
framework.
}
