% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_metrics.R
\name{eigen_error}
\alias{eigen_error}
\title{Eigenvalue Error}
\usage{
eigen_error(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{Reference covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
A numeric value giving the mean absolute eigenvalue error.
}
\description{
Computes the mean absolute difference between the eigenvalues of a reference
covariance matrix and an estimated covariance matrix.
}
