% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_protect.R
\name{empty_matrix}
\alias{empty_matrix}
\title{Create an empty matrix}
\usage{
empty_matrix(n)
}
\arguments{
\item{n}{Integer. Number of rows in the matrix.}
}
\value{
A numeric matrix with \code{n} rows and 0 columns.
}
\description{
Creates an empty numeric matrix with a specified number of rows and
zero columns.
}
\keyword{internal}
