% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_metrics.R
\name{frobenius_norm}
\alias{frobenius_norm}
\title{Frobenius Norm Error}
\usage{
frobenius_norm(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{Reference covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
A numeric value giving the Frobenius norm error.
}
\description{
Computes the Frobenius norm of the difference between a reference covariance
matrix and an estimated covariance matrix.
}
