% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_method.R
\name{mom_calculation}
\alias{mom_calculation}
\alias{mom_calculation.univariate}
\title{Estimate Empirical-Bayes Hyperparameters}
\usage{
mom_calculation(type, ...)

\method{mom_calculation}{univariate}(type, gamma_hat, delta_hat, ...)
}
\arguments{
\item{type}{Object used for S3 method dispatch.}

\item{...}{Additional arguments passed to the corresponding method.}

\item{gamma_hat}{Vector of additive batch-effect estimates.}

\item{delta_hat}{Vector of multiplicative batch-effect estimates.}
}
\value{
Estimated empirical-Bayes hyperparameters.

A list containing method-of-moments estimates for the location and
scale hyperparameters.
}
\description{
S3 generic for estimating method-of-moments hyperparameters used in
empirical-Bayes shrinkage.
}
