% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_metrics.R
\name{mse_cov}
\alias{mse_cov}
\title{Mean Squared Covariance Error}
\usage{
mse_cov(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{Reference covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
A numeric value giving the mean squared covariance error.
}
\description{
Computes the mean squared element-wise difference between a reference
covariance matrix and an estimated covariance matrix.
}
