% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_cca_cor}
\alias{plot_cca_cor}
\title{Plot Canonical Correlations}
\usage{
plot_cca_cor(cca_batch, cca_cov)
}
\arguments{
\item{cca_batch}{A result from \code{\link[=cca_diagnostic]{cca_diagnostic()}} for batch or technical
variables.}

\item{cca_cov}{A result from \code{\link[=cca_diagnostic]{cca_diagnostic()}} for biological covariates.}
}
\value{
A \code{ggplot} object.
}
\description{
Compares canonical correlations between surrogate variables and batch or
technical variables versus biological covariates.
}
