% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{print.latentcombat_inference}
\alias{print.latentcombat_inference}
\title{Print LatentComBat Inference Results}
\usage{
\method{print}{latentcombat_inference}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=harm_inference]{harm_inference()}}.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print LatentComBat Inference Results
}
