% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{print.summary_latentcombat_fit}
\alias{print.summary_latentcombat_fit}
\title{Print a LatentComBat Summary}
\usage{
\method{print}{summary_latentcombat_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_latentcombat_fit"}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Prints a compact summary of a \code{"summary_latentcombat_fit"} object.
}
