% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_metrics.R
\name{spectral_norm}
\alias{spectral_norm}
\title{Spectral Norm Error}
\usage{
spectral_norm(Sigma, Sigma_hat)
}
\arguments{
\item{Sigma}{Reference covariance matrix.}

\item{Sigma_hat}{Estimated covariance matrix.}
}
\value{
A numeric value giving the spectral norm error.
}
\description{
Computes the spectral norm of the difference between a reference covariance
matrix and an estimated covariance matrix.
}
