% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{tsne_plot}
\alias{tsne_plot}
\title{Plot t-SNE Embedding by Batch}
\usage{
tsne_plot(tsne_prep_result, ellipse = TRUE)
}
\arguments{
\item{tsne_prep_result}{A data frame returned by \code{\link[=tsne_prep]{tsne_prep()}}.}

\item{ellipse}{Logical indicating whether to display batch-specific
68-percent concentration ellipses. Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a scatter plot of a two-dimensional t-SNE embedding, colored by
batch or site.
}
